Update README.md
This lab is meant to serve as a tutorial, introducing you to the embedded system you will utilize in this course and the software development environment for it. It will also serve as a review of some material from ECE 2049.

To complete all labs in this course, you will need to submit your source for that lab, and demonstrate the lab in operation to one of the course staff (a TA or tutor). You will be graded on the completion of the steps in this lab document. After you have received a signoff, follow the instructions at the end of the document to submit your source code.

Objectives

Install software: Code Composer Studio and TI RTOS

Set up a Code Composer Studio project

Output to the 128×128 LCD display

Configure general purpose timers

Configure the interrupt controller

Read and debounce user buttons